app/requirements.txt
fastapi==0.116.1
python-gedcom==1.0.0
python-multipart==0.0.20
uvicorn==0.35.0
